defect_detection_project_FINAL/defect_detection/requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=9.0.0

# Data Science / EDA
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Explainability
grad-cam>=1.4.8

# Deployment
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
timm>=0.9.0
